% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateData.R
\name{aggregateData}
\alias{aggregateData}
\title{Aggregate validation data for further analysis}
\usage{
aggregateData(data, aggregation = "month", na.rm = TRUE)
}
\arguments{
\item{data}{Object returned from getData (class validationData)}

\item{aggregation}{Time aggregation. Values for "month" and "year" are automatically converted. Custom aggregation by specifying a date format (i.e. used in format.Date) which is subsequently converted to a date if possible (i.e. used in as.Date)}

\item{na.rm}{Handles how to deal with NA's}
}
\value{
An object containing aggregated values for nloc, nsim, ntime, time, observations, simulations and the requested attributes
}
\description{
Aggregate validation data for further analysis
}
