% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLocationsFromBoundingBox.R
\name{getLocationsFromBoundingBox}
\alias{getLocationsFromBoundingBox}
\title{Get the location of stations based on the mask of the observation file and a bounding box}
\usage{
getLocationsFromBoundingBox(obsFile, boundingBox, obsVar = "GRDC_number")
}
\arguments{
\item{obsFile}{Path to the observation file.}

\item{boundingBox}{Vector containing minimum longitude, maximum longitude, minimum latitude, maximum latitude (in that order)}

\item{obsVar}{bservation variable to be used as mask. If the variable is NA, it is assumed  the cell does not contain data. Variable dimensions > 2 only use the first slice of redundand dimensions. Defaults to GRDC_number.}
}
\value{
List of station locations containing lon/lat values in a vector.
}
\description{
Get the location of stations based on the mask of the observation file and a bounding box
}
