% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLocationsFromSimulations.R
\name{getLocationsFromSimulations}
\alias{getLocationsFromSimulations}
\title{Get the location of stations based on the mask of the observation file and the mask of the simulation file(s)}
\usage{
getLocationsFromSimulations(obsFile, simFile, simVar = "OUT_DISCHARGE",
  obsVar = "area_observed")
}
\arguments{
\item{obsFile}{Path to the observation file.}

\item{simVar}{Simulation variable to be used as mask. If the variable is NA, it is assumed the cell does not contain data. Variable dimensions > 2 only use the first slice of redundand dimensions. Defaults to VIC discharge variable.}

\item{obsVar}{Observation variable to be used as mask. If the variable is NA, it is assumed  the cell does not contain data. Variable dimensions > 2 only use the first slice of redundand dimensions. Defaults to GRDC number.}

\item{simFiles}{Path to the simulation file.}
}
\value{
List of station locations containing lon/lat values in a vector.
}
\description{
Get the location of stations based on the mask of the observation file and the mask of the simulation file(s)
}
\examples{
obsFile = system.file("data", "obsSampleRhine.nc", package = "dischargeValidation", mustWork = T)
simFiles = system.file("data", "simSampleRhine.nc", package = "dischargeValidation", mustWork = T)

locations = getLocationsFromSimulations(obsFile = obsFile, simFile = simFiles)
}
