% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSimValues.R
\name{getSimValues}
\alias{getSimValues}
\title{getSimValues}
\usage{
getSimValues(simFile, location, simOrigin, simVar, values)
}
\arguments{
\item{simFile}{}

\item{location}{}

\item{simOrigin}{}

\item{simVar}{}

\item{values}{}
}
\value{
data.frame with the observation time and values and the simulation values (added as a column)
}
\description{
Gather simulation values from NetCDF file based on the observation time domain
}
