% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBias.R
\name{plotBias}
\alias{plotBias}
\title{Make scatter plot of temporal mean simulated versus observed discharge.}
\usage{
plotBias(data_all_loc, perc_pr = 90, log_plot = F, power_symb = 0.25,
  spec_dis = T, fyear = 0, lyear = 0, title_plot = " ", ...)
}
\arguments{
\item{data_all_loc}{Object with the data (observations, simulations and meta data) obtained by a call with getData.}

\item{perc_pr}{Minimum percentage of all time slots during the period from fyear to lyear with both a valid observation AND a valid simulation. i.e. with a valid data pair. Stations with less than perc_pr valid data pairs are not plotted. Mean values of observations and simulations will be computed for all time slots with a valid data pair. Default: 90.}

\item{log_plot}{Plot with logarithmic (T) or linear (F) axes. Default: linear axes (F).}

\item{power_symb}{This parameter affects the relative size of the symbols in the plot. The area of the circles representing individual stations are proportional to the area of the catchment upstream from the station (according to the meta data of the observations) to the power power_symb. Default: 0.25.}

\item{spec_dis}{Plot specific (T) or absolute amount (F) of discharge. Specific discharge (in mm/d) is equal to absolute amount of discharge (in m3/s) divided by the area of the catchment upstream from the station (for the observations) or area of the catchment according to the routing network (for simulations; this includes the area of the grid cell in which the station is located).}

\item{fyear}{First year of the period considered. Default: first year of data_all_loc.}

\item{lyear}{Last year of the period considered. Default: last year of data_all_loc.}

\item{title_plot}{Title for the plot. Default: fyear-lyear, e.g. 1981-2000.}

\item{...}{Parameters accepted by the R function plot can be added.}
}
\value{
A data frame with each row representing a station. Columns contain respectively values for the temporal mean difference between simulation and observation, for the temporal mean of the observations and the temporal mean of the simulations, for the river and station name, for the latitude and longitude of the station, for the area of the catchment according to the meta data of the observations and the area of the catchment of the model routing network and for the GRDC number. Rows are ranked according to the relative difference between simulation and observation.
}
\description{
Make scatter plot of temporal mean simulated versus observed discharge.
}
