% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{getData}
\usage{
getData(obsFile, simFiles, locations, simOrigins = rep("0001-01-01",
  length(simFiles)), simVars = rep("OUT_DISCHARGE", length(simFiles)),
  obsOrigin = "1900-01-01", obsVar = "dis", attributes = c("all"))
}
\arguments{
\item{obsFile}{}

\item{simFiles}{}

\item{locations}{}

\item{simOrigins}{}

\item{simVars}{}

\item{obsOrigin}{}

\item{obsVar}{}

\item{attributes}{}
}
\value{
list containing validationData classes for each location
}
\description{
Get observation and simulation data and attributes at specified locations
}
