% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDatum.R
\name{getDatum}
\alias{getDatum}
\title{getDatum}
\usage{
getDatum(obsFile, simFiles, location, simOrigins, simVars, obsOrigin,
  obsVar, attributes)
}
\arguments{
\item{attributes}{}
}
\value{
validationData classes for the location
}
\description{
Get observation and simulation data and attributes at specified location
}
