% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLocationsFromSimulations.R
\name{getLocationsFromSimulations}
\alias{getLocationsFromSimulations}
\title{getLocationsFromSimulations}
\usage{
getLocationsFromSimulations(obsFile, simFile, simVar = "OUT_DISCHARGE",
  obsVar = "GRDC_number")
}
\arguments{
\item{obsVar}{}
}
\value{
list of locations (lat/lon)
}
\description{
Get station locations based on the simulation domain
}
