% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validationDataClass.R
\name{isValidationData}
\alias{isValidationData}
\title{Checks wether an object is of type "validationData"}
\usage{
isValidationData(x)
}
\arguments{
\item{x}{The object}
}
\value{
true or false
}
\description{
Checks wether an object is of type "validationData"
}
\examples{
obsFile = system.file("data", "obsSampleRhine.nc", package = "dischargeValidation", mustWork = T)
simFiles = system.file("data", "simSampleRhine.nc", package = "dischargeValidation", mustWork = T)
locations = getLocationsFromBoundingBox(obsFile = obsFile, boundingBox = c(6.75,10.25,47.75,51.25))
data = getData(obsFile = obsFile, simFiles = simFiles, location = locations[1:4])

isValidationData(data)
}
