% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validationDataClass.R
\name{isValidationData}
\alias{isValidationData}
\title{Checks wether an object is of type "validationData"}
\usage{
isValidationData(x)
}
\arguments{
\item{x}{The object}
}
\value{
true or false
}
\description{
Checks wether an object is of type "validationData"
}
