% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validationDataClass.R
\name{isValidationData}
\alias{isValidationData}
\title{isValidationData}
\usage{
isValidationData(x)
}
\arguments{
\item{x}{}
}
\value{
true or false
}
\description{
check if valid validationData class
}
